refactor(ops): mutualise le parsing d'arguments dans lib/args
getArg/hasFlag étaient copiés dans 4 scripts (verify-dashboard-stats garde sa convention espacée).
